Rearrange ginkgo suite helpers
