+++ pkg/test/ginkgo/test_suite.go
@@ -8,33 +8,37 @@ import (
 	"strings"
 	"time"
 
+	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/ginkgo/v2/types"
+
+	"k8s.io/apimachinery/pkg/util/errors"
+
+	"github.com/openshift/origin/test/extended/util/annotate/generated"
 )
 
-type testCase struct {
-	name      string
-	spec      types.TestSpec
-	locations []types.CodeLocation
-	apigroups []string
+func testsForSuite() ([]*testCase, error) {
+	var tests []*testCase
+	var errs []error
 
-	// identifies which tests can be run in parallel (ginkgo runs suites linearly)
-	testExclusion string
-	// specific timeout for the current test. When set, it overrides the current
-	// suite timeout
-	testTimeout time.Duration
+	// Don't build the tree multiple times, it results in multiple initing of tests
+	if !ginkgo.GetSuite().InPhaseBuildTree() {
+		ginkgo.GetSuite().BuildTree()
+	}
 
-	start           time.Time
-	end             time.Time
-	duration        time.Duration
-	testOutputBytes []byte
-
-	flake    bool
-	failed   bool
-	skipped  bool
-	success  bool
-	timedOut bool
-
-	previous *testCase
+	ginkgo.GetSuite().WalkTests(func(name string, spec types.TestSpec) {
+		if append, ok := generated.Annotations[name]; ok {
+			spec.AppendText(append)
+		}
+		tc, err := newTestCaseFromGinkgoSpec(spec)
+		if err != nil {
+			errs = append(errs, err)
+		}
+		tests = append(tests, tc)
+	})
+	if len(errs) > 0 {
+		return nil, errors.NewAggregate(errs)
+	}
+	return tests, nil
 }
 
 var re = regexp.MustCompile(`.*\[Timeout:(.[^\]]*)\]`)
@@ -65,6 +69,32 @@ func newTestCaseFromGinkgoSpec(spec types.TestSpec) (*testCase, error) {
 	}
 
 	return tc, nil
+}
+
+type testCase struct {
+	name      string
+	spec      types.TestSpec
+	locations []types.CodeLocation
+	apigroups []string
+
+	// identifies which tests can be run in parallel (ginkgo runs suites linearly)
+	testExclusion string
+	// specific timeout for the current test. When set, it overrides the current
+	// suite timeout
+	testTimeout time.Duration
+
+	start           time.Time
+	end             time.Time
+	duration        time.Duration
+	testOutputBytes []byte
+
+	flake    bool
+	failed   bool
+	skipped  bool
+	success  bool
+	timedOut bool
+
+	previous *testCase
 }
 
 func (t *testCase) Retry() *testCase {
